move test classes from Tests namespace
