+++ composer.json
@@ -29,7 +29,7 @@
 	},
 	"autoload-dev": {
 		"psr-4": {
-			"Kucera\\Monolog\\Tests\\": "tests/Monolog/"
+			"Kucera\\Monolog\\": "tests/Monolog/"
 		}
 	},
 	"extra": {
