+++ tests/Monolog/TestCase.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Kucera\Monolog\Tests;
+namespace Kucera\Monolog;
 
 abstract class TestCase extends \PHPUnit_Framework_TestCase
 {
